Updated profile handling

+++ JSCocoa Delegate/JSCocoaDelegate.m
@@ -41,8 +41,8 @@
 	}
 	
 	// register all arguments in JS context
-	for (NSUInteger i = 0; i < [arguments count]; i++) {
-		[argNames addObject:[NSString stringWithFormat:@"__objcArg%ld", i]];
+	for (int i = 0; i < [arguments count]; i++) {
+		[argNames addObject:[NSString stringWithFormat:@"__objcArg%d", i]];
 		[ctx setObject:[arguments objectAtIndex:i] withName:[argNames lastObject]];
 	}
